+++ seantis/dir/events/tests/test_import.py
@@ -213,37 +213,38 @@ class TestImport(IntegrationTestCase):
             ids = ['event1', 'event2', 'event3', 'event4',
                    'event5', 'event6', 'event7', 'event8']
             events = from_ids(ids[:4])
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 4)
             imported = [i.getObject().source_id for i in self.catalog.query()]
             self.assertEquals(ids[:4], imported)
 
             # Import with limit
             events = from_ids(ids[4:])
-            imports = importer.fetch_one('source', fetch, limit=2)
+            imports, deleted = importer.fetch_one('source', fetch, limit=2)
             self.assertEquals(imports, 2)
             self.assertEquals(len(self.catalog.query()), 6)
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 2)
             self.assertEquals(len(self.catalog.query()), 8)
 
             # Force reimport
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 0)
             self.assertEquals(len(self.catalog.query()), 8)
-            imports = importer.fetch_one('source', fetch, reimport=True)
+            imports, deleted = importer.fetch_one('source', fetch,
+                                                  reimport=True)
             self.assertEquals(imports, 4)
             self.assertEquals(len(self.catalog.query()), 8)
 
             # Reimport updated events
             events = from_ids(ids)
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 8)
             self.assertEquals(len(self.catalog.query()), 8)
 
             # Test import of given source IDs only
             events = from_ids(ids)
-            imports = importer.fetch_one(
+            imports, deleted = importer.fetch_one(
                 'source', fetch, source_ids=ids[2:6]
             )
             self.assertEquals(imports, 4)
@@ -273,7 +274,7 @@ class TestImport(IntegrationTestCase):
                 cat1=set(), cat2=set(['cat2-1', 'cat2-2', 'cat2-3'])
             ))
 
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 3)
             self.assertTrue('cat1-1' in self.directory.cat1_suggestions)
             self.assertTrue('cat1-2' in self.directory.cat1_suggestions)
@@ -321,7 +322,7 @@ class TestImport(IntegrationTestCase):
             # event['attachment_1'] =
             # event['attachment_2'] =
 
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             imported = self.catalog.query()
             self.assertEquals(imports, 1)
             self.assertEquals(len(imported), 1)
@@ -363,8 +364,9 @@ class TestImport(IntegrationTestCase):
                                                   fetch_id='fetch_id'))
             events.append(self.create_fetch_entry(source_id='source_id_2',
                                                   fetch_id='fetch_id'))
-            imports = importer.fetch_one('source', fetch)
+            imports, deleted = importer.fetch_one('source', fetch)
             self.assertEquals(imports, 3)
+            self.assertEquals(deleted, 0)
             self.assertEquals(len(self.directory.keys()), 3)
 
             # Second import
@@ -373,8 +375,10 @@ class TestImport(IntegrationTestCase):
             events.append(self.create_fetch_entry(source_id='source_id_3',
                                                   fetch_id='fetch_id'))
 
-            imports = importer.fetch_one('source', fetch, autoremove=True)
+            imports, deleted = importer.fetch_one('source', fetch,
+                                                  autoremove=True)
             self.assertEquals(imports, 1)
+            self.assertEquals(deleted, 2)
 
             ids = [item.source_id for id, item in self.directory.objectItems()]
             self.assertEquals(len(ids), 2)
@@ -389,7 +393,7 @@ class TestImport(IntegrationTestCase):
             # Import event
             importer = ExternalEventImporter(self.directory)
             event = self.create_fetch_entry(source_id='s', fetch_id='f')
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             self.assertEquals(imports, 1)
 
             # Hide event
@@ -401,9 +405,9 @@ class TestImport(IntegrationTestCase):
             hidden.hide()
 
             # Re-import event
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             self.assertEquals(imports, 0)
-            imports = importer.fetch_one(
+            imports, deleted = importer.fetch_one(
                 'source', lambda: [event], reimport=True
             )
             self.assertEquals(imports, 1)
@@ -427,7 +431,7 @@ class TestImport(IntegrationTestCase):
             # Import event
             importer = ExternalEventImporter(self.directory)
             event = self.create_fetch_entry(source_id='s', fetch_id='f')
-            imports = importer.fetch_one('source', lambda: [event])
+            imports, deleted = importer.fetch_one('source', lambda: [event])
             self.assertEquals(imports, 1)
 
             # Add own event
@@ -598,7 +602,7 @@ class TestImport(IntegrationTestCase):
             "long_description": "h\u00e4nsel",
             "cat1": ["cat12", "cat11"], "cat2": ["cat21"],
             "event_url": "http://www.event.ch",
-            "timezone": "UTC",
+            "timezone": "Europe/Zurich",
             "start": "2014-01-15T00:00:00+00:00",
             "end": "2014-01-15T23:59:59+00:00",
             "whole_day": true,
@@ -627,8 +631,8 @@ class TestImport(IntegrationTestCase):
             "cat1": ["cat13", "cat14"], "cat2": ["cat21", "cat22", "cat23"],
             "event_url": null,
             "timezone": "UTC",
-            "start": "2014-01-19T17:00:00+00:00",
-            "end": "2014-01-19T18:00:00+00:00",
+            "start": "2014-01-19T17:00:00+02:00",
+            "end": "2014-01-19T18:00:00+02:00",
             "whole_day": false, "recurrence": null,
             "locality": null, "street": null, "housenumber": null,
             "zipcode": null, "town": null,
@@ -679,7 +683,7 @@ class TestImport(IntegrationTestCase):
         self.assertEquals(events[0]['contact_phone'], '+12 (3) 45 678 90 12')
         self.assertEquals(events[0]['latitude'], '1.0')
         self.assertEquals(events[0]['longitude'], '2.0')
-        self.assertEquals(events[0]['timezone'], default_timezone())
+        self.assertEquals(events[0]['timezone'], 'Europe/Zurich')
         self.assertEquals(events[0]['start'], datetime(2014, 1, 15, 0, 0,
                                                        tzinfo=pytz.UTC))
         self.assertEquals(events[0]['end'], datetime(2014, 1, 15, 23, 59, 59,
@@ -698,9 +702,10 @@ class TestImport(IntegrationTestCase):
         )
         self.assertEquals(events[1]['latitude'], None)
         self.assertEquals(events[1]['longitude'], None)
-        self.assertEquals(events[1]['start'], datetime(2014, 1, 19, 17, 0,
+        self.assertEquals(events[1]['timezone'], 'UTC')
+        self.assertEquals(events[1]['start'], datetime(2014, 1, 19, 15, 0,
                                                        tzinfo=pytz.UTC))
-        self.assertEquals(events[1]['end'], datetime(2014, 1, 19, 18, 0,
+        self.assertEquals(events[1]['end'], datetime(2014, 1, 19, 16, 0,
                                                      tzinfo=pytz.UTC))
         self.assertEquals(events[1]['whole_day'], False)
         self.assertEquals(events[1]['cat1'], set(['cat13', 'cat14']))